refactor adding new comment
