+++ src/App.css
@@ -66,9 +66,36 @@ button {
   border-radius: 100%;
 }
 
+.form-header {
+  display: flex;
+  align-items: center;
+}
+
 .cancel-button {
   cursor: pointer;
   padding: 5px;
+  margin-left: auto;
+  margin-right: 0px;
+  display: flex;
+}
+
+.loader {
+  border: 5px solid #f3f3f3;
+  border-top: 5px solid dodgerblue;
+  border-radius: 50%;
+  width: 12px;
+  height: 12px;
+  animation: spin 2s linear infinite;
+  margin-left: auto;
+}
+
+@keyframes spin {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
 }
 
 .green {
@@ -116,10 +143,36 @@ footer > section {
   text-transform: capitalize;
 }
 
-form {
+.comment {
+  padding: 5px;
+  margin-right: 20px;
+  margin-left: 20px;
+}
+
+textarea {
+  font-size: 1rem;
+  padding: 5px;
+  height: 80px;
+  position: relative;
+}
+
+.comment-label {
   display: flex;
-  max-width: 750px;
+}
+
+.form-buttons,
+.form-header {
+  display: flex;
   align-items: flex-end;
+  margin: 5px;
+  flex: 1;
+}
+
+.comment-input {
+  width: 100%;
+  box-sizing: border-box;
+  border: 1px solid rgb(213, 212, 212);
+  border-radius: 8px;
 }
 
 .form-button {
@@ -130,25 +183,14 @@ form {
   display: inline-block;
   text-transform: capitalize;
   font-size: 1rem;
+  margin-left: auto;
+  margin-right: 0px;
+  display: flex;
 }
 
-textarea {
-  font-size: 1rem;
-  margin: 5px;
-  padding: 5px;
-  width: 100%;
-  height: 100px;
-  display: block;
-}
-
-.input-area {
-  display: block;
-  flex-grow: 4;
-  margin: 20px;
-}
-
-.form-buttons {
-  margin: 20px;
+.neutral {
+  background-color: dodgerblue;
+  margin-right: auto;
 }
 
 *:disabled {
